git-phab: Pass --keep-cr to git-am to avoid breaking on CRLF patches
If a patch (and the file it is patching) use CRLF line endings, git-am
will not work, as it normalises the line endings in the file (but not
the patch) before applying. Pass --keep-cr to git-am to avoid this — it
will now apply the patches, but will spew warnings about trailing
whitespace (the CR characters), which seems fair.

Signed-off-by: Philip Withnall <philip.withnall@collabora.co.uk>
Reviewed-by: Thibault Saunier <tsaunier@gnome.org>
Differential Revision: https://phabricator.freedesktop.org/D1266