Use branches from staging repo to set task branches
No reason to use some other repo if a staging repo is avalaible.

Also if no remote URI is set on the task we won't be able
to later `git phab apply TXXX` because it is currently impossible
to retrieve the list of differential associated to a task.

Here we simply set the URI to the last published diff which
contains the whole branch as expected.

The following error was happenning:

    Can not apply revisions from a task if no 'remote branch' has been set for it.
    INFO: You need to find what revisions are associated with the tasks and apply them.

Differential Revision: https://phabricator.freedesktop.org/D1672